Remove 'delay' from 'validateOperation'
If the contract needs to enforce different delay for different selectors
or method parameters, it will not be able to expose the
'scheduleDelayedOps(batch)' method as it will not enforce individual
method's requirements.

This commit fixes the test for this by adding an expensive check on all
batch operations not to contain 'HelloWorld', which is not how this
should be done.