Implement and test static 'deployNewVaultFactory' in the Factory Interactor
There is no need to deploy the Factory from the Foundation itself. On
the other hand, I want to avoid duplicationg this code.

Also, separate the 'Factory Interactor' and 'Factroy Interactor
Promises', as async JS functions must be wrapped with Kotlin's 'suspend
await' method.

Also, this means we need to include the '/src/js/generated/' folder in
the NPM package (no need to check in to git).