fix unknownParticipants in SimpleWallet
in the UI, the wallet failed to to decode guardians, and reported
unknownParticipants - even though the same flow passed OK in the system
test.
Root cause: the SimpleManager.guardianAddress was not set. it was a
constructor parameter, and was only set in the TestEnvironment, so it
wasn't clear it was missing in the "real UI app.

- removed this param. it can (and should) be extracted from the
'backend.getAddresses()'
- added async _init() to SimpleWallet, for one-time async
initializations (which can't be in the constructor)
- added nonNull() check function, which verifies that classes receive
values for specific params. Later, we might add stricter type-checking
on params...